Added the Actual and removed the test component
